feat(flake): Enable SpiceTify for Spotify
This commit message follows the specified format and provides a concise description of the changes made in the code diff. It uses the "feat" type to indicate that this is a new feature addition, specifically enabling SpiceTify for Spotify within the flake.nix configuration. The message is under 72 characters as required.